feat: move MTLS package to pkg instead of internal
